requirements.txt
pandas>=1.5.0
numpy>=1.21.0
setuptools
apache-airflow-providers-google==10.10.1
sqlalchemy
psycopg2-binary
apache-airflow-providers-postgres>=5.0.0
scikit-learn>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
imbalanced-learn